Update task_ivar_consensus to consider multi-contig references
Natively, ivar consensus is not capable of working with more than
one reference name at a time. This is a problem for analyzing
viruses with multiple segments. The proposed solution aims to run
ivar consensus on each contig aligned to the reference individually and
concatenate them together at the end. These changes shouldn't impact
previous implementations of this task.