Introduce skani_status output variable
This variable will be used downstream to determine if skani was
successful in finding/selecting a reference genome.

Previously, we've implemented a catch for highly fragmented de novo
assemblies where all contigs are concatenated together in a last
ditch effort to generate a suitable reference genome.

However, sometimes this "supercontig" is still not adequate (too small)
for skani to find an appropriate reference genome. Added an additional
check for cases like this and it also reports the length of the combined contig
lengths.