Fix consensus when no reads align to segmented contig
Mpileup with the -aa flag will output mpileups for all contigs and will
report all reference positions regardless of quality/mapping coverage.
We need to have this flag enabled to keep track of all positions, but we
must also include the -r regions flag to specify the contig we are
currently iterating through.