Update find command and include debug statements
Turns out the find command is nondeterministic and the order in which
the model files are returned can vary from system to system. Now we
specifically look for the directory with the least nested *.minfo file
to report as the `--mdir` parameter. Should fix any issues with finding
the correct flu model files.