[VADR] Refactor and Add Support for Additional Viruses (#882)
* Rework VADR model handling

Now all VADR models are stored in a requester pays bucket and accessed
during runtime. Includes support for sc2, mpxv, wnv, flu-a, flu-b,
rsv-a, rsv-b. New VADR support for measles, mumps, and rubella.

* Update documentation

* Update standalone VADR workflow

* Update workflows referencing VADR in organism_parameters

All of these need to be added in order to manually override the VADR
model across each workflow.

* Add new docker image and update workflows referencing VADR

Need to make sure that VADR is only called for applicable organisms
across all workflows.

* Fixing several minor workfow syntax issues

* Fix outputs for org-specific vadr defaults

* Add reference genomes and genome len for new organisms

* Update conditional organisms to trigger VADR task

* Updated WNV vadr_options and fixed order of commands

* Update I/O

* Make model file extraction more explicit

* Troubleshooting model file extraction

I think the issue could be the variable expansion ${model_files_location}.

* Add VADR parameters to TheiaCoV_ClearLabs

* Remove unnecessary inputs

These are already exposed at the organism_parameter workflow level.

* Fix wonky indentation

* Add VADR model file to miniwdl/cromwell test inputs

* Update md5sums

* Update docs

* Increase memory for flu

* Update find command and include debug statements

Turns out the find command is nondeterministic and the order in which
the model files are returned can vary from system to system. Now we
specifically look for the directory with the least nested *.minfo file
to report as the `--mdir` parameter. Should fix any issues with finding
the correct flu model files.

* Return flu VADR memory to default (8GB)

* Update md5sums

* Fix docs

* Update I/O

* Update VADR conditionals in morgana_magic

* Update docs and workflow diagram

* update docs

* Add new taxon IDs to theiaviral taxa-specific docs

* Add mumps and rubella to standardized_organism output