Removed `bases_to_crop` and redundant default parameters
The `bases_to_crop` variable is only used in the TBProfiler_tNGS
workflow and it's supposed to remove X number of bases from the start
of a read, but it also unintentionally sets the "max read length" to the
avg_readlength.

Just going to remove this and give the user the option to set any
trimming parameters they want to use.