Restrict primer trim length dynamically
This variable performs `ktrim` on both ends of a read, examining only
the outermost X number of bases. Important to note that this does not
include the number of bases your primer reference sequence has. So it
needs to be at least the min length of your primer sequence. Will make
clear in docs.